feat: kotlin coroutines, driver class, more optimized hopefully
change gobilda pinpoint driver to kotlin, use coroutines to ensure that they are run on a different thread, however use a mutex to ensure that there is thread saftey.

scope is defined as the next available thread.

abstract: this change makes reads more asynchronous by running them on a coroutine thread.